Make forum-post-cooldown use auto-follow-topic's logic
